requirements.txt
Flask==2.0.1
newsapi-python==0.2.6
requests==2.26.0
gunicorn==20.1.0
Werkzeug==2.0.3
python-dotenv==1.0.0
Flask-Caching==1.10.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
